+++ src/main/java/roomescape/service/command/PaymentCommandService.java
@@ -1,6 +1,7 @@
 package roomescape.service.command;
 
 import org.springframework.stereotype.Service;
+import roomescape.client.PaymentClient;
 import roomescape.domain.payment.Payment;
 import roomescape.dto.payment.PaymentResponseDto;
 import roomescape.dto.reservation.PaymentConfirmDto;
@@ -8,14 +9,14 @@ import roomescape.dto.reservation.PaymentConfirmDto;
 @Service
 public class PaymentCommandService {
 
-    private final PaymentService paymentService;
+    private final PaymentClient paymentClient;
 
-    public PaymentCommandService(final PaymentService paymentService) {
-        this.paymentService = paymentService;
+    public PaymentCommandService(final PaymentClient paymentClient) {
+        this.paymentClient = paymentClient;
     }
 
     public PaymentResponseDto confirmPayment(PaymentConfirmDto requestDto) {
-        Payment payment = paymentService.confirmPayment(requestDto);
+        Payment payment = paymentClient.confirmPayment(requestDto);
         return new PaymentResponseDto(payment.getOrderId(), payment.getTotalAmount());
     }
 }
